Fix a few >= issues, process papers where reference start not found (rare), add more constants
Pruning named references, the reference is assumed to be OK if the years of
consecutive references are not equal, and the lines between the reference is
above 2 (rather than >= 2), this gives better results

Fixed issue where if the first reference was on the references start line (i.e.
if the references start point was not found and the sequence start was used),
the reference numbers would be invalid because the first ref would not be
included, i.e. refs would start from 2 (silver et al, game of go)

When the reference start point is not found, just use the start of the first
sequence as the reference starting point. This probably works for dotted refs,
but not others. It seems to be a rare case.

Add some constants for minimum chars for a reference, and maximum lines. Below
minimum the reference is ignored, above maximum lines it is truncated.